Changes the way we move an object
First of all, this will solve the circular reference issue between
@uplayground@ and @umovingobject@ (through @ubouncingobject@)

Also, this will (hopefully) fix the slowdown issue around objects (although
this issue was never actually seen, it was belived to happen. See comment
in the old @aMovingObject.move()@ function). Note that discretization of the
path is now at the discretion of the manager (i.e. @oPlayground@), and hence
no longer handled by @aMovingObject@.